backend/requirements.txt
fastapi
pyjwt
passlib
python-jose
uvicorn
python-dotenv
cryptography